+++ helpers.js
@@ -23,4 +23,17 @@ const urlForUsers = function (userID, urlDatabase) {
 	}
 	return newUrlDatabase;
 };
-module.exports = { findUserByEmail, urlForUsers };
+
+function generateRandomString() {
+	var result = "";
+	var characters =
+		"ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
+	var charactersLength = characters.length;
+	for (var i = 0; i < 6; i++) {
+		result += characters.charAt(Math.floor(Math.random() * charactersLength));
+	}
+
+	return result;
+}
+
+module.exports = { findUserByEmail, urlForUsers, generateRandomString };
